requirements.txt
# Core LLM & Orchestration
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Web Search
tavily-python>=0.5.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0

# Frontend
streamlit>=1.40.0
httpx>=0.27.0

# Vector Store (RAG)
faiss-cpu>=1.8.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Linting
ruff>=0.8.0
